move components in shop page to components
