+++ src/components/shop/buy/index.ts
@@ -2,3 +2,4 @@ export * from './LikeTabPost'
 export * from './LikeTabPostList'
 export * from './OfferTabPost'
 export * from './OfferTabPostList'
+export * from './panel'
