+++ src/components/shop/index.ts
@@ -1,3 +1,4 @@
+export * from './layout'
 export * from './sale'
 export * from './buy'
 export * from './review'
