+++ src/components/shop/review/index.ts
@@ -1,2 +1,3 @@
 export * from './ReviewTabPost'
 export * from './ReviewTabPostList'
+export * from './panel'
