+++ src/components/shop/sale/index.ts
@@ -1,2 +1,3 @@
+export * from './panel'
 export * from './SaleTabPost'
 export * from './SaleTabPostList'
