+++ src/components/shop/sale/panel/styled.ts
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 import type { ColorKeys } from '@offer-ui/react'
 import { Divider as DividerComponent } from '@offer-ui/react'
-import { Tabs, Tab as TabComponent } from '@components'
+import { Tabs, Tab as TabComponent } from '@components/common'
 
 const SearchOptionsWrapper = styled.div`
   ${({ theme }): string => `
